+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 ### Additions
 * `iguanatweaksreborn:explosion_damage_multiplier` NBT tag for entities to increase/decrease explosion damage
+* Villagers now heal every ~10 seconds
 
 ### Changes
 * Upgrade Aquatic beachgrass no longer drops seeds